Pokemon Data Visualization
Trying to organize Pokemon by type, and ideally, generation once I
figure this out. I started out with cycling only through one type to see
if it works. It doesn't, and I think it's because the program isn't
using the final value of 'normalNum' inside the 'types' function. Still
trying to figure out how to work around that.